Updated PDF Parser with functions to dynamically handle PDF files
